requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.27.1
playwright>=1.40.0
python-dotenv==1.0.1
lxml==5.3.0
pydantic==2.10.5
celery==5.4.0
redis==5.2.0
flower==2.0.1
sqlalchemy>=2.0.36
pymysql==1.1.0
cryptography>=43.0.0
markdown==3.7
pygments==2.18.0